GIX-1440: Move env-var to utils
